pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "house-search"
version = "0.1.0"
description = "A semantic search system for real estate properties using vision-language models"
requires-python = ">=3.12"

dependencies = [
    # Environment Configuration
    "python-dotenv>=1.0.0",

    # Natural Language Processing
    "nltk>=3.8.0",

    # Text Embedding
    "sentence-transformers>=5.1.2",

    # Vector Database
    "pinecone>=3.0.0",

    # Image Processing
    "Pillow>=10.0.0",

    # Web Interface
    "streamlit>=1.28.0",
]

[project.optional-dependencies]
# Pipeline/Batch Processing dependencies (for image-to-text and embedding pipelines)
pipeline = [
    # Core dependencies
    "numpy>=1.24.0",
    "pandas>=2.0.0",

    # Deep Learning Framework
    "torch>=2.1.0",
    "torchvision>=0.15.0",

    # Vision-Language Model
    "transformers>=4.57.3",
    "huggingface-hub>=0.36.0",
    "accelerate>=0.24.0",
    "qwen-vl-utils>=0.0.2",

    # Data Storage
    "pyarrow>=13.0.0",
    "fastparquet>=2023.10.0",

    # Progress bars
    "tqdm>=4.66.0",
]

dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[tool.setuptools]
packages = ["app", "common", "image_to_text_pipeline", "text_to_embedding_pipeline", "batch_jobs"]

[tool.black]
line-length = 100
target-version = ["py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
    \.venv
    | \.git
    | __pycache__
    | build
    | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_gitignore = true
known_first_party = ["app", "batch_jobs", "common", "image_to_text_pipeline", "text_to_embedding_pipeline", "scripts"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
exclude = [
    "^\\.venv/",
    "^build/",
    "^dist/",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov-report=term-missing"
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "SIM"]
ignore = ["E501"]
exclude = [
    ".venv",
    ".git",
    "__pycache__",
    "build",
    "dist",
]
